pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "momentum-scanner"
version = "1.0.0"
description = "Short-term momentum scanner for UK and US equities"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Scanner Team"}
]
keywords = ["trading", "scanner", "momentum", "technical-analysis", "stocks"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "rich>=13.5.0",
    "tabulate>=0.9.0",
    "python-dotenv>=1.0.0",
    "tzdata>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.0.290",
]

[project.scripts]
scanner-cli = "scanner.modes.cli:main"
scanner-worker = "scanner.modes.worker:main"

[project.urls]
Homepage = "https://github.com/awaistechnologist/momentum-scanner-py"
Repository = "https://github.com/awaistechnologist/momentum-scanner-py"

[tool.setuptools.packages.find]
where = ["."]
include = ["scanner*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
